backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
psycopg2-binary
celery
redis
requests
pyjwt
djangorestframework-simplejwt
bcrypt
pytz
stripe
gunicorn
whitenoise
python-dotenv
cryptography
email-validator
drf-yasg
django-ratelimit
django-extensions
google-auth-oauthlib
google-auth-httplib2
graphe
graphene-django
channels
channels-redis
django-filter
scikit-learn
numpy
pandas
django-db-connection-pool
django-cachalot
